Move trungify-related commands to cogs/image_manipulation.py
